requirements.txt
streamlit
pytesseract
pillow
openpyxl
pandas
opencv-python